+++ lib/models/budgetData.dart
@@ -1,19 +1,19 @@
 class BudgetData{
-  int food,entertainment,travel,snacks,fuel,bills,shopping,health,others;
+  double food,entertainment,travel,snacks,fuel,bills,shopping,health,others;
   BudgetData(){
     food=0;entertainment=0;travel=0;snacks=0;fuel=0;bills=0;shopping=0;health=0;others=0;
   }
 }
 
 class WeeklyData{
-  int sun,mon,tue,wed,thu,fri,sat;
+  double sun,mon,tue,wed,thu,fri,sat;
   WeeklyData(){
     sun=0;mon=0;tue=0;wed=0;thu=0;fri=0;sat=0;
   }
 }
 
 class YearlyData{
-  int jan,feb,mar,apr,may,jun,jul,aug,sep,oct,nov,dec;
+  double jan,feb,mar,apr,may,jun,jul,aug,sep,oct,nov,dec;
   YearlyData(){
     jan=0;feb=0;mar=0;apr=0;may=0;jun=0;jul=0;aug=0;sep=0;oct=0;nov=0;dec=0;
   }
